implemented heap and memory allocation

+++ include/types.h
@@ -18,7 +18,7 @@ typedef uint64_t uintptr_t;
 typedef unsigned long int uintmax_t;
 
 typedef unsigned long size_t;
-typedef long long ssize_t;
+typedef long ssize_t;
 #else
 #define "undefined architecture"
 #endif
